refactor: move utility related files to new `src/utils` directory
